Pre-release of version 0.2.0.  Added file uploads, changed file writing to plugin-like system, added FTP file writing support, fixed lots of bugs. See the ChangeLog for more details.

+++ themes/skepticats/templates/blogentry_tpl.php
@@ -3,16 +3,12 @@
 	<h2><a href="<?php echo $PERMALINK; ?>"><?php echo $SUBJECT; ?></a></h2>
 	<ul>
 		<li class="blogdate">Posted <?php echo $POSTDATE; ?></li>
-		<?php
-			if (false) {
-		?>
+<?php	if (check_login() ) { ?>
 		<li class="blogadmin">
 			<a href="<?php echo $POSTEDIT; ?>">Edit</a>
 			<a href="<?php echo $POSTDELETE; ?>">Delete</a>
 		</li>
-		<?php 
-			}
-		?>
+<?php } ?>
 	</ul>
 </div>
 <div class="blogentrybody">
